requirements_proj.txt
docling>=2.0.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
langchain-community>=0.0.10
chromadb>=0.4.18
pydantic>=2.0.0
python-dotenv>=1.0.0 